feat(api): move /v2/users routes to new structure
