--- packages/appeals-service-api/src/routes/appeal-users.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const { userGet } = require('../controllers/user');
-const { userPost } = require('../controllers/appeal-user');
-const router = express.Router();
-
-router.get('/:email', userGet);
-router.post('/', userPost);
-module.exports = router;
