+++ packages/appeals-service-api/src/routes/index.js
@@ -20,7 +20,7 @@ const lpaDashboardAppealsRouter = require('./appeals-case-data');
 const documentMetadataRouter = require('./documentMetadata');
 const responsesRouter = require('./responses');
 const listedBuildingRouter = require('./listed-building');
-const appealUsersRouter = require('./appeal-users');
+const { routes: v2Routes } = require('./v2');
 
 router.use('/api/v1/appeals', appealsRouter);
 router.use('/api/v1/back-office', backOfficeRouter);
@@ -35,6 +35,10 @@ router.use('/api/v1/appeals-case-data', lpaDashboardAppealsRouter);
 router.use('/api/v1/document-meta-data', documentMetadataRouter);
 router.use('/api/v1/responses', responsesRouter);
 router.use('/api/v1/listed-buildings', listedBuildingRouter);
-router.use('/api/v2/users', appealUsersRouter);
+
+// v2 routes loaded from the file structure
+for (const [url, handler] of Object.entries(v2Routes)) {
+	router.use(`/api/v2/${url}`, handler);
+}
 
 module.exports = router;
