+++ packages/appeals-service-api/src/routes/v2/users/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const { userGet, userPost } = require('./controller');
+const router = express.Router();
+
+router.get('/:email', userGet);
+router.post('/', userPost);
+
+module.exports = { router };
